requirements.txt
Flask
requests
python-dotenv
Flask-Login
Flask-Mail
psycopg2-binary
Flask-SQLALchemy
oauthlib
pyOpenSSL
